rename folder tgb_modules to modules for merge into tgb
